Fix undefined behavior in avalanche.cpp
Summary:
In the MMIX LCG calculation, the right side calculation underlying type
is signed. This causes the undefined behavior sanitizer to detect a
signed integer overflow.
This diff sets the constants to `unsigned long long` to enforce the
underlying type to be unsigned and get rid of this error.

Test Plan:
With the UBSAN activated, run
  test_bitcoin -t avalanche_tests

Reviewers: #bitcoin_abc, deadalnix, markblundeberg

Reviewed By: #bitcoin_abc, deadalnix, markblundeberg

Subscribers: markblundeberg

Differential Revision: https://reviews.bitcoinabc.org/D3074

+++ src/avalanche.cpp
@@ -95,7 +95,8 @@ bool VoteRecord::addNodeToQuorum(NodeId nodeid) {
     }
 
     // MMIX Linear Congruent Generator.
-    const uint64_t r1 = 6364136223846793005 * nodeid + 1442695040888963407;
+    const uint64_t r1 =
+        6364136223846793005 * uint64_t(nodeid) + 1442695040888963407;
     // Fibonacci hashing.
     const uint64_t r2 = 11400714819323198485ull * (nodeid ^ seed);
     // Combine and extract hash.
